requirements.txt
pandas
matplotlib
dataspyre
